requirements.txt
fastapi
uvicorn
jinja2
python-multipart
transformers
torch
google-generativeai
python-dotenv